apps/tasks/requirements.txt
boto3==1.20.52
botocore==1.23.52
requests==2.24.0
psycopg2-binary==2.9.3